requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.15
requests==2.31.0
pyOpenSSL==23.2.0 
Flask-Login==0.6.2
python-dotenv==1.0.0
oauthlib==3.2.2
psycopg2-binary==2.9.6
flask-bcrypt==1.0.1
pyjwt==2.7.0
pymongo==4.4.1
pytest==7.4.0
docker==6.1.3
Flask-Session==0.5.0
